Move pydantic_compat out of typing_compat
