+++ docs/user/tutorials/create-a-device.rst
@@ -118,7 +118,7 @@ if the device requires any adapters to control it externally.
 
 .. code-block:: python
 
-    from tickit.utils.compat.typing_compat import pydantic_dataclass
+    from tickit.utils.compat.pydantic_compat import pydantic_dataclass
 
     from tickit.core.components.component import Component, ComponentConfig
     from tickit.core.components.device_simulation import DeviceSimulation
