+++ examples/devices/amplifier.py
@@ -9,7 +9,7 @@ from tickit.core.components.device_simulation import DeviceSimulation
 from tickit.core.device import Device, DeviceUpdate
 from tickit.core.typedefs import SimTime
 from tickit.utils.byte_format import ByteFormat
-from tickit.utils.compat.typing_compat import pydantic_dataclass
+from tickit.utils.compat.pydantic_compat import pydantic_dataclass
 
 
 class AmplifierDevice(Device):
