+++ examples/devices/zeromq_push_device.py
@@ -9,7 +9,7 @@ from tickit.adapters.zeromq.push_adapter import (
 from tickit.core.components.component import Component, ComponentConfig
 from tickit.core.components.device_simulation import DeviceSimulation
 from tickit.devices.iobox import IoBoxDevice
-from tickit.utils.compat.typing_compat import pydantic_dataclass
+from tickit.utils.compat.pydantic_compat import pydantic_dataclass
 
 
 class IoBoxZeroMqAdapter(ZeroMqPushAdapter):
