+++ src/tickit/adapters/zeromq/push_adapter.py
@@ -17,7 +17,7 @@ import zmq
 
 from tickit.core.adapter import Adapter, RaiseInterrupt
 from tickit.core.device import Device
-from tickit.utils.compat.typing_compat import BaseModel
+from tickit.utils.compat.pydantic_compat import BaseModel
 
 LOGGER = logging.getLogger(__name__)
