+++ src/tickit/utils/configuration/loading.py
@@ -3,7 +3,7 @@ from typing import List
 import yaml
 
 from tickit.core.components.component import ComponentConfig
-from tickit.utils.compat.typing_compat import parse_obj_as
+from tickit.utils.compat.pydantic_compat import parse_obj_as
 
 
 def read_configs(config_path) -> List[ComponentConfig]:
