+++ tests/utils/configuration/test_loading.py
@@ -6,7 +6,7 @@ from mock.mock import mock_open
 
 from tickit.core.components.component import ComponentConfig
 from tickit.core.typedefs import ComponentID, ComponentPort, PortID
-from tickit.utils.compat.typing_compat import pydantic_dataclass
+from tickit.utils.compat.pydantic_compat import pydantic_dataclass
 from tickit.utils.configuration.loading import read_configs
